+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 0.1.1
+## 0.2.0
 
 * FluentForms - do_shortcode in Terms & Conditions
 * Disable editor fullscreen mode
@@ -10,7 +10,14 @@
   * Fix quotes
   * Reveal Reusable posts in menu
   * Disable directory search
+  * Disable layout
+  * Disable render global styles
+* FontAwesome new icons:
+  * Download
+* WP helpers
+  * Get blog ID & title
+* Theme - disable jQuery migrate only on frontend
 
-## 0.1.0
+## 0.1.0 (2022-01-26)
 
 * Initial release (jsmprf version).